refactor: components folder restructuring
