+++ src/components/sections/TeamPortrait.tsx
